pyproject.toml
[project]
name = "bigpotatopizzahey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygithub>=2.8.1",
]
